Removing step to archive anonymous model on triton deployment (#1997)
* Remove archiving model step

No need to archive an anonymous model registered by the system. Eventually, this action will trigger an error given it is a stage transition not allowed.

* Removing archive model step

No need to archive an anonymous model registered by the system. Eventually, this action will trigger an error given it is a stage transition not allowed.

* Update online-endpoints-triton-cc.ipynb

* Update online-endpoints-triton.ipynb